+++ frontend/src/components/HelloWorld.js
@@ -1,0 +1,9 @@
+// src/components/HelloWorld.js
+
+import React from "react";
+
+function HelloWorld() {
+  return <h2>Hello, World!</h2>;
+}
+
+export default HelloWorld;
