+++ frontend/src/components/LoginPage.js
@@ -1,0 +1,57 @@
+// src/components/LoginPage.js
+
+import React, { useState } from "react";
+import { useNavigate } from "react-router-dom";
+
+function LoginPage() {
+  const [username, setUsername] = useState("");
+  const [password, setPassword] = useState("");
+  const [message, setMessage] = useState("");
+  const navigate = useNavigate();
+
+  const handleRegister = async () => {
+    const response = await fetch("http://localhost:5000/register", {
+      method: "POST",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify({ username, password }),
+    });
+    const data = await response.json();
+    setMessage(data.message);
+  };
+
+  const handleLogin = async () => {
+    const response = await fetch("http://localhost:5000/login", {
+      method: "POST",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify({ username, password }),
+    });
+    const data = await response.json();
+    setMessage(data.message);
+    if (data.success) {
+      navigate("/hello");
+    }
+  };
+
+  return (
+    <div className="container">
+      <h1>Login or Sign Up</h1>
+      <input
+        type="text"
+        placeholder="Username"
+        value={username}
+        onChange={(e) => setUsername(e.target.value)}
+      />
+      <input
+        type="password"
+        placeholder="Password"
+        value={password}
+        onChange={(e) => setPassword(e.target.value)}
+      />
+      <button onClick={handleLogin}>Login</button>
+      <button onClick={handleRegister}>Register</button>
+      {message && <p>{message}</p>}
+    </div>
+  );
+}
+
+export default LoginPage;
